Unit test moved to Optimize/AsyncCSS + revisit code changes
